feat: add create class feature for teachers
This commit adds a new feature for teachers to create a new class using a POST request to the API. The ClassesController has been updated to include a new action method named "PostClass" that allows a teacher to create a new class. The teacher's user ID is extracted from the JWT token and used to assign the teacher as the owner of the class. The class data is validated and saved to the database using the ClassRepository. The new class data is returned in the response body along with a 201 Created status code.