feat(post): add ability to upload multiple resources
Added a new feature to the post module that allows users to upload multiple resources when creating a new post. The resources are stored in the database and can be viewed on the post detail page. This feature should improve the user experience by making it easier to share multiple resources in a single post.